content/html/content/test/forms/test_submit_invalid_file.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=702949
-->
<head>
  <meta charset="utf-8">
  <title>Test invalid file submission</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=702949">Mozilla Bug 702949</a>
<p id="display"></p>
<div id="content" style="display: none">
  <form action='submit_invalid_file.sjs' method='post' target='result'
        enctype='multipart/form-data'>
    <input type='file' name='file'>
  </form>
  <iframe name='result'></iframe>
</div>
<pre id="test">
</pre>
<script type="application/javascript">
  /*
   * Test invalid file submission by submitting a file that has been deleted
   * from the file system before the form has been submitted.
   * The form submission triggers a sjs file that shows its output in a frame.
   * That means the test might time out if it fails.
   */

  SimpleTest.waitForExplicitFinish();
  addLoadEvent(function() {
    var FileUtils = SpecialPowers.Cu.import("resource://gre/modules/FileUtils.jsm").FileUtils;

    var i = document.getElementsByTagName('input')[0];

    var file = FileUtils.getFile("TmpD", ["testfile"]);
    file.createUnique(SpecialPowers.Ci.nsIFile.NORMAL_FILE_TYPE, 0644);

    SpecialPowers.wrap(i).value = file.path;
    file.remove(/* recursive = */ false);

    document.getElementsByName('result')[0].addEventListener('load', function() {
      is(window.frames[0].document.body.textContent, "SUCCESS");
      SimpleTest.finish();
    });
    document.forms[0].submit();
  });
</script>
</body>
</html>
